+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Use scientific Decimal rendering for inline literals and require Ch `~> 0.8.2` https://github.com/plausible/ecto_ch/pull/276
+
 ## 0.9.2 (2026-05-06)
 
 - Require Ch `~> 0.8.1` and move `function_exported?(Decimal, :to_string, 3)` to compile-time https://github.com/plausible/ecto_ch/pull/275